+++ src/App.vue
@@ -1,6 +1,6 @@
 <template>
-	<demo-file-input />
+	<demo-input />
 </template>
 
 <script lang="ts" setup>
-	import DemoFileInput from "./demos/DemoFileInput.vue";</script>
+	import DemoInput from "./demos/DemoInput.vue";</script>